#1162: consider timeout also in outer loop of UpdateManager
